purchases_customer/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
pillow==11.3.0
django-filter==23.3
setuptools==65.0.0
drf-spectacular==0.28.0
pandas==2.3.2
openpyxl==3.1.5